requirements.txt
streamlit==1.27.2
pandas==1.5.3
numpy==1.23.5
python-docx==0.8.11
openpyxl==3.1.2
matplotlib==3.7.1
seaborn==0.12.2